refactor(dashboard): RFC-0050 PR-1 — cost-dashboard.ts ownership split
Mechanical split per RFC-0050 §4.1. Three pure modules extracted from
cost-dashboard.ts; render code stays. All public exports preserved
via re-export so external imports (`runtime-panel.ts`,
`cost-dashboard.test.ts`) need no change.

Files:
- dashboard/src/components/cost/cost-types.ts (34 LoC)
  Type aliases + guards (CostView/CostFocus/AuditFocus/ViewMode +
  isCostView/isCostFocus/isAuditFocus + viewModeForCostFocus +
  auditRouteParams).
- dashboard/src/components/cost/cost-formatters.ts (14 LoC)
  Pure formatters: formatTokens + severityClass.
- dashboard/src/components/cost/audit-summarizer.ts (84 LoC)
  AuditActorSummary + AuditKindSummary interfaces + private helpers
  (latestTs, mostFrequentKind) + exported severityBuckets +
  summarizeAuditActors + summarizeAuditKinds.

Cost-dashboard.ts shrank 1442 → 1352 LoC (-90). Net +42 LoC (~3%) in
the dashboard tree from the import/re-export glue, well under
RFC-0050 §5's ~10% allowance.

Why these three files (RFC-0050 §3 criteria check):

1. Plurality of distinct domains: ✓ types, formatters, audit aggregation
   are independently coherent and reusable. Already named in the RFC.
2. No cross-domain shared mutable state: ✓ all three new files are
   pure — no signals, no module-level state, no side effects.
3. Domain count ≥ 3: ✓ exactly three.
4. No path-based reach into private helpers: ✓ verified — only public
   exports are imported from this file by external callers
   (runtime-panel.ts: CostDashboard, CostView; cost-dashboard.test.ts:
   8 names already in the re-export block).
5. No public-export rename: ✓ every name preserved verbatim.

severityBuckets was promoted from private to exported because the
render at line ~899 calls it directly (AuditSummaryBoard StatTile).
Same function, same body — only the visibility changed.

NOT included in this PR (deferred): connector-status.ts split. On
re-verification it FAILS RFC-0050 §3 criterion 1 — zero per-connector
branches, all four connectors share one render parameterized by
`connectorId: string`. The "split per-connector" idea in RFC-0050
§4.1 was based on data-row counts (3 lookup tables of 4 rows each =
~30 LoC), not implementation structure (~1650 LoC of shared render
code). Splitting would move 30 LoC out and leave the rest intact —
exactly the §2 anti-pattern (mechanical line-count split with no
ownership distribution). That candidate is dropped from RFC-0050;
follow-up PR will amend §4.1 to reflect this finding.
